+++ lib/src/utils/parsing/parseRoutes.ts
@@ -1,16 +1,16 @@
-import { getRoutesTreeChain } from './getRoutesTreeChain';
-import { uniqueByIdAndNestingDepth } from './uniqueByIdAndNestingDepth';
-
 export function parseRoutes(routes: RouteMatchData[], url: string) {
-  const allMatched: RouteMatchData[] = [];
-  routes.forEach(
-    (route) =>
-      route.regexpPath.test(url) &&
-      allMatched.push(...getRoutesTreeChain(routes, route))
-  );
-  const unique = uniqueByIdAndNestingDepth(allMatched);
-  // if (!unique.length) {
-  //   console.error('[Easyroute] No routes matched')
-  // }
-  return unique;
+  const usedIds: string[] = [];
+  const usedDepths: number[] = [];
+
+  return routes.filter((route) => {
+    if (
+      usedIds.includes(route.id) ||
+      usedDepths.includes(route.nestingDepth) ||
+      !route.regexpPath.test(url)
+    )
+      return false;
+    usedIds.push(route.id as string);
+    usedDepths.push(route.nestingDepth as number);
+    return true;
+  });
 }
