+++ lib/src/utils/path/constructUrl.ts
@@ -3,7 +3,5 @@ import { deleteFirstSlash } from './deleteFirstSlash';
 
 export function constructUrl(url: string, base: string) {
   if (!base || url.includes(base)) return url;
-  const strippedBase = deleteLastSlash(base);
-  const strippedUrl = deleteFirstSlash(url);
-  return `/${strippedBase}/${strippedUrl}`;
+  return `/${deleteLastSlash(base)}/${deleteFirstSlash(url)}`;
 }
